// ClientApp/src/Types/ICompanies.ts
export default interface ICompanies {
  companies_companyID: string;
  companies_name: string;
  companies_uniqueCode: string;
  companies_phonenumber: string;
  companies_email: string;
  companies_isPremium: boolean;
  companies_isActive: boolean;

  companies_websiteUrlLink: string;
  companies_advertMessage: string;
  companies_imagesLink: string;

  companies_physicalAddress: string;
}